requirements.txt
aiogram
python-dotenv
aiosqlite
aiohttp
